+++ packages/solana/lib/solana.dart
@@ -23,3 +23,4 @@ typedef TransactionId = String;
 typedef Wallet = Ed25519HDKeyPair;
 
 const int lamportsPerSol = 1000000000;
+const int solDecimalPlaces = 9;
