+++ packages/solana/lib/solana_pay.dart
@@ -1,0 +1,1 @@
+export 'src/solana_pay/solana_pay.dart';
