+++ packages/solana/lib/src/solana_pay/constants.dart
@@ -1,2 +1,2 @@
 const String solanaPayScheme = 'solana';
-const int solDecimalPlaces = 9;
+
