+++ packages/solana/lib/src/solana_pay/exceptions.dart
