+++ packages/solana/lib/src/solana_pay/solana_pay_request.freezed.dart
