--- packages/solana_pay/CHANGELOG.md
@@ -1,3 +1,0 @@
-## 1.0.0
-
-- Initial version.
