--- packages/solana_pay/Makefile
@@ -1,1 +1,0 @@
-include ../../Makefile
