--- packages/solana_pay/analysis_options.yaml
@@ -1,1 +1,0 @@
-include: package:mews_pedantic/analysis_options.yaml
