--- packages/solana_pay/example/main.dart
@@ -1,3 +1,0 @@
-void main() {
-  // TODO(KB): Add example
-}
