--- packages/solana_pay/lib/solana_pay.dart
@@ -1,5 +1,0 @@
-library solana_pay;
-
-export 'src/exceptions.dart';
-export 'src/solana_client_ext.dart';
-export 'src/solana_pay_request.dart';
