--- packages/solana_pay/pubspec.yaml
@@ -1,19 +1,0 @@
-name: solana_pay
-description: Dart library for Solana Pay protocol support.
-version: 0.1.0
-# homepage: https://www.example.com
-
-environment:
-  sdk: '>=2.16.1 <3.0.0'
-
-
-dependencies:
-  decimal: ^2.1.0
-  freezed_annotation: ^1.1.0
-  solana: ^0.22.0-dev.2
-
-dev_dependencies:
-  build_runner: ^2.1.8
-  freezed: ^1.1.1
-  mews_pedantic: ^0.4.1
-  test: ^1.16.0
